requirements.txt
Django>=1.8.2
djangorestframework
markdown
django-filter
